networkd: Add new property to monitor addresses changes
Currently, in DBUS API, there is no way to know the current IP address
of a configured interface, or to be notified in case of changes.

This adds a new property to get the current IP address of the
corresponding link, to the /org/freedesktop/network1/link interface. It
also adapts the networkd-address.c module to send DBUS signal when this
address is changed.

Signed-off-by: Romain Perier <romain.perier@collabora.com>
Signed-off-by: Fabien Lahoudere <fabien.lahoudere@collabora.co.uk>