networkd: Emit PropertyChanged from address_add() when address already exists
Currently, when a network interface is not plugged but already configured
during bootup, the change of the 'Addresses' property will be signaled,
early and the interface will be configured without carrier gained. Then
if the iface is plugged, carrier will be gained but PropertyChanged
won't be re-emitted a second time. This is an issue, because 'Addresses'
property is supposed to be changed and to be signaled each time the
interface is plugged/unplugged.

This commit fixes the issue, it ensures that Addresses PropertyChanged
is emitted for a configured interface  when carrier is gained.

Signed-off-by: Romain Perier <romain.perier@collabora.com>
Signed-off-by: Fabien Lahoudere <fabien.lahoudere@collabora.co.uk>